fix build

+++ setup.py
@@ -83,26 +83,31 @@ def get_requirements() -> List[str]:
 
 
 def prepare_submodules() -> None:
-    def _run_cmd(args: str) -> None:
+    def _run_cmd(args: str, check: bool = True) -> None:
         cmds = args.split(" ")
-        subprocess.run(cmds)
+        returned = subprocess.run(cmds, stderr=subprocess.STDOUT)
+        if check:
+            returned.check_returncode()
+        elif returned.returncode != 0:
+            logger.warning("Run %s error, please check!" % args)
 
     old_dir = os.getcwd()
-
     os.chdir(ROOT_DIR)
 
-    # Fetch source codes.
-    _run_cmd("git submodule update --init vllm_mindspore/msadapter")
-    os.chdir(get_path("vllm_mindspore", "msadapter"))
-    _run_cmd("git reset --hard HEAD")
+    msadapter_path = Path() / "vllm_mindspore" / "msadapter"
+    if not any(msadapter_path.iterdir()):
+        # Fetch source codes.
+        _run_cmd("git submodule update --init vllm_mindspore/msadapter", check=False)
+        os.chdir(get_path("vllm_mindspore", "msadapter"))
+        _run_cmd("git reset --hard HEAD")
 
-    patch_dir = Path(ROOT_DIR) / "patch" / "msadapter"
-    for p in patch_dir.glob("*msadapter*"):
-        _run_cmd("git apply {}".format(p))
+        patch_dir = Path(ROOT_DIR) / "patch" / "msadapter"
+        for p in patch_dir.glob("*.patch"):
+            _run_cmd("git apply {}".format(p))
 
-    # Add __init__.py for packing.
-    _run_cmd("touch __init__.py")
-    _run_cmd("touch mindtorch/__init__.py")
+        # Add __init__.py for packing.
+        _run_cmd("touch __init__.py")
+        _run_cmd("touch mindtorch/__init__.py")
 
     os.chdir(old_dir)
